+++ morpheus-graphql-core/src/Data/Morpheus/Types/Internal/AST.hs
@@ -168,6 +168,7 @@ import Data.Morpheus.Ext.OrdMap
 import Data.Morpheus.Types.Internal.AST.Base
 import Data.Morpheus.Types.Internal.AST.DirectiveLocation (DirectiveLocation (..))
 import Data.Morpheus.Types.Internal.AST.Fields
+import Data.Morpheus.Types.Internal.AST.InpuUnion
 import Data.Morpheus.Types.Internal.AST.Selection
 import Data.Morpheus.Types.Internal.AST.Stage
 import Data.Morpheus.Types.Internal.AST.TH
@@ -175,7 +176,7 @@ import Data.Morpheus.Types.Internal.AST.TypeCategory
 import Data.Morpheus.Types.Internal.AST.TypeSystem
 import Data.Morpheus.Types.Internal.AST.Value
 import Language.Haskell.TH.Syntax (Lift)
-import Prelude (Show)
+import Prelude (Either (..), Show)
 
 type Variables = HashMap FieldName ResolvedValue
